pathlib is better

+++ main.py
@@ -1,55 +1,56 @@
 #!/usr/bin/env python
-import os
 import shutil
 from argparse import ArgumentParser
 from datetime import datetime
+from pathlib import Path
 
 from PIL import Image
 from PIL.ExifTags import TAGS
 
 
-def get_image_date_taken(file_path):
+def get_image_date_taken(filepath: Path):
     """Extracts the date taken from image metadata."""
     result = None
     try:
-        image = Image.open(file_path)
-        exif_data = image._getexif() if hasattr(image, "_getexit") else dict()
+        # Attempt to read exif with pillow
+        image = Image.open(filepath)
+        exif_data = image._getexif() if hasattr(image, "_getexif") else dict()
         for tag_id, value in exif_data.items():
             tag_name = TAGS.get(tag_id, tag_id)
             if tag_name == "DateTimeOriginal":
                 result = datetime.strptime(value, "%Y:%m:%d %H:%M:%S")
                 break
         else:
-            raise ValueError(f"{file_path} has no exif data.")
-    except Exception as err:
-        timestamp = os.path.getmtime(file_path)
+            raise ValueError(f"{filepath} has no exif data.")
+    except Exception:
+        timestamp = filepath.stat().st_mtime
         result = datetime.fromtimestamp(timestamp)
     print(".", end="")
     return result
 
 
-def organize_images_by_date(source, destination, use_copy):
+def organize_images_by_date(source: Path, destination: Path, use_copy: bool):
     """Organizes images into folders by year/month/day."""
-    print("scanning", source)
-
     file_counter = 0
-    for root, _, files in os.walk(source):
-        for file in files:
-            src_filepath = os.path.join(root, file)
-
+    seen = set()
+    for src_filepath in source.rglob("*"):
+        if src_filepath.is_file():
             try:
                 date_taken = get_image_date_taken(src_filepath)
-
                 year = date_taken.strftime("%Y")
                 month = date_taken.strftime("%m - %B")
-                day_meta = "{day} - {name}".format(
-                    day=date_taken.strftime("%d"), name=os.path.basename(root)
+                day_meta = "{day:02} - {name}".format(
+                    day=date_taken.day, name=src_filepath.parent.name.strip()
                 )
-                dst_folder = os.path.join(destination, year, month, day_meta)
-                if not os.path.exists(dst_folder):
-                    os.makedirs(dst_folder)
 
-                dst_filepath = os.path.join(dst_folder, file)
+                # prepare target folder
+                dst_folder = destination / year / month / day_meta
+                if dst_folder not in seen:
+                    dst_folder.mkdir(parents=True, exist_ok=True)
+                    seen.add(dst_folder)
+
+                # move/copy target file
+                dst_filepath = dst_folder / src_filepath.name
                 file_counter += 1
                 if use_copy:
                     shutil.copy(src_filepath, dst_filepath)
@@ -59,7 +60,7 @@ def organize_images_by_date(source, destination, use_copy):
             except Exception as err:
                 print(f"Cannot process {src_filepath}, reason: {err}")
 
-    return file_counter
+    return file_counter, len(seen)
 
 
 if __name__ == "__main__":
@@ -69,13 +70,15 @@ if __name__ == "__main__":
     parser.add_argument("destination")
     args = parser.parse_args()
 
-    print(args.source, "to", args.destination)
-    if not os.path.isdir(args.source):
-        parser.error(f"{args.source} folder does not exist.")
-    elif not os.path.isdir(args.destination):
-        parser.error(f"{args.destination} folder does not exists.")
+    source = Path(args.source).resolve()
+    destination = Path(args.destination).resolve()
+    print("Reorganizing", source, "to", destination)
 
-    source = os.path.realpath(args.source)
-    file_count = organize_images_by_date(source, args.destination, args.copy)
+    if not source.is_dir():
+        parser.error(f"{source} folder does not exist.")
+    elif not destination.is_dir():
+        parser.error(f"{destination} folder does not exist.")
+
+    files, folders = organize_images_by_date(source, destination, args.copy)
     print()
-    print(f"Organized {file_count} files.")
+    print(f"Organized {files} files in {folders} folders.")
